I2C and SPI moved to helpers subdir
